RISC-V: Add embeddedramdisk shell command
It adds a command which can mount a FV section as a ramdisk.
To be able to execute anything from there, the SecurityStub checker needs to be
disabled.
Otherwise it will postpone execution of any APP or DXE after DXE phase, if it
is not contained in the FW image.